refactor(extraction): split 1372-line module into a package (#100)
extraction.py used to hold ExtractionSchema, ExtractionResult,
ClaudeExtractor, prompt constants, and the legacy dealer / spam
helpers in a single file. Each PR touching one concern had to be
reviewed against five.

Split, behavior-neutral. ``git mv`` preserves blame on the bulk.

extraction/                       (line counts after split)
  __init__.py        86           back-compat re-exports + PEP 562
  schema.py         127           ExtractionSchema dataclass
  spam.py            57           parse_bool / contains_dealer_text /
                                  seller_looks_like_dealer + legacy
                                  underscore aliases
  result.py         141           ExtractionResult
  extractor.py     1085           ClaudeExtractor + prompt constants
                                  (was 1372; -287 lines moved out)

Backward compat preserved
- ``from mantis_agent.extraction import ExtractionSchema`` etc. still
  work via the package __init__ re-exports.
- ``from mantis_agent.extraction import _parse_bool`` (and the two
  other underscore helpers) still work — aliased in spam.py.
- ``from mantis_agent.extraction import DEALER_TEXT_INDICATORS`` still
  works via the package's PEP 562 __getattr__ with the same
  DeprecationWarning that was on the old single-file module.

Tests / lint
- pytest tests/ -q -> 370 passed (same as after #104; pure refactor).
- ruff check on the new package: clean.

Out of scope (acceptance criterion #1 still partial)
- ``extractor.py`` is 1085 lines — over the 800-line target the issue
  set. Reducing it further requires splitting ClaudeExtractor itself
  (extract / extract_multi / find_listing_content_control / find_all)
  into mixins or sub-classes; that's an architectural decision worth
  its own PR.
- Same split treatment for baseten_server.py and modal_runtime.py is
  the next two PRs against this issue.

Closes #100 (partial — see "Out of scope" above).
